Add BaseHelper that will now carry the server methods.
